Add Convolution operator benchmarks
