+++ benchmark/opperf/opperf.py
@@ -26,7 +26,8 @@ from benchmark.opperf.nd_operations.binary_operators import run_mx_binary_broadc
 from benchmark.opperf.nd_operations.gemm_operators import run_gemm_operators_benchmarks
 from benchmark.opperf.nd_operations.random_sampling_operators import run_mx_random_sampling_operators_benchmarks
 from benchmark.opperf.nd_operations.nn_activation_operators import run_activation_operators_benchmarks
-from benchmark.opperf.nd_operations.nn_pooling_operators import run_pooling_operators_benchmarks
+from benchmark.opperf.nd_operations.nn_conv_operators import run_pooling_operators_benchmarks, \
+    run_convolution_operators_benchmarks
 
 from benchmark.opperf.utils.common_utils import merge_map_list, save_to_file
 
@@ -65,6 +66,9 @@ def run_all_mxnet_operator_benchmarks(ctx=mx.cpu(), dtype='float32'):
 
     # Run all Pooling operations benchmarks with default input values
     mxnet_operator_benchmark_results.append(run_pooling_operators_benchmarks(ctx=ctx, dtype=dtype))
+
+    # Run all Convolution operations benchmarks with default input values
+    mxnet_operator_benchmark_results.append(run_convolution_operators_benchmarks(ctx=ctx, dtype=dtype))
 
     # ****************************** PREPARE FINAL RESULTS ********************************
     final_benchmark_result_map = merge_map_list(mxnet_operator_benchmark_results)
